requirements.txt
numpy==2.4.0
Pillow==12.0.0
scikit_learn==1.8.0
tensorboardx==2.6.4
torch==2.9.0
torchvision==0.24.1
